Use new unwinder for offline in libbacktrace.
libunwindstack changes:
- Add a load_bias field in MapInfo so that it can be loaded offline,
  and also so it can be cached.
- Add an Add function the Maps class so that it's possible to manually
  create a map.
- Remove the OfflineMaps class since I haven't found a reason for this to
  exist.
- Add a pointer to the gnu debugdata compressed section in the interface
  itself and modify the step path to try debug_frame, then eh_frame, then
  gnu_debugdata. This way arm can add exidx as the last step behind
  gnu_debugdata. Add an offline test for this feature.
- Fix x86_64_ucontext_t since it was a different size on 32 bit and 64 bit
  systems.

libbbacktrace changes:
- Completely rewrite the BacktraceOffline class to use the new unwinder.
- Modify the test data to save ucontext_t data instead of unw_context data.
- Convert the previous tests from unw_context data to ucontext_t data.

Bug: 65682279

Test: New unit tests pass in libunwindstack.
Test: All offline tests continue to pass.

+++ debuggerd/libdebuggerd/utility.cpp
@@ -200,7 +200,7 @@ void dump_memory(log_t* log, unwindstack::Memory* memory, uint64_t addr, const c
   size_t total_bytes = start + bytes;
   for (size_t line = 0; line < MEMORY_BYTES_TO_DUMP / MEMORY_BYTES_PER_LINE; line++) {
     std::string logline;
-    android::base::StringAppendF(&logline, "    %" PRIPTR, addr);
+    android::base::StringAppendF(&logline, "    %" PRIPTR, static_cast<uint64_t>(addr));
 
     addr += MEMORY_BYTES_PER_LINE;
     std::string ascii;
